adds overloads to aggregator::determineTimestamp
For the stream aligner and transformer to work, access to a timestamp in
the message is needed. By default a member variable called "time" is
assumed to exist, which is not the case in the objectDetectionMessages.
Therefore, overloads to determineTimestamp are provided which correctly
access the timestamp in the header of the messages.